simplify health check code
